Legal-Check-main/requirements.txt
streamlit>=1.35.0
groq>=0.9.0
pymupdf>=1.24.0
python-docx>=1.1.0
python-dotenv>=1.0.0